honksquad: refactor: split SharedCarryingSystem into partials (#854)
Behavior-preserving extraction of the 604-line SharedCarryingSystem into
five focused partial-class files, keeping it one system:

- SharedCarryingSystem.cs        - Initialize + Carry/Drop/CanCarry core
                                   (plus the shared carry do-after, pull-block
                                   and speed/movement hooks).
- SharedCarryingSystem.Verbs.cs  - verb providers + interrupt do-after.
- SharedCarryingSystem.DragDrop.cs - drag-drop adapter.
- SharedCarryingSystem.AutoDrop.cs - the ~7 pure Drop() delegations collapse
                                   into a shared DropFromInterrupt(uid) reached
                                   through generic Subscribe{Carried,Carrier}Drop
                                   {,Ref} helpers; only the conditional handlers
                                   (mob-state, parent-change) keep bespoke bodies.
- SharedCarryingSystem.Cleanup.cs - OnBeingCarriedShutdown broken into
                                   TeardownCarrier / Reparent / RestorePosture.

Every subscription, ref/by-value dispatch and teardown ordering is preserved;
subscriptions stay in Initialize via InitializeXxx/SubscribeXxx helpers so
HONK0009 is satisfied. Public Carry/Drop/InterruptCarry API is unchanged and
still covered by Content.IntegrationTests/Tests/RussStation/Carrying.
